added webp support package to project
This package should enable support for webp assets. Note that there was an error when adding the package because of a duplicate .dll file shared between this package and the Yarn Spinner package (System.Runtime.Compiler.Services.Unsafe.dll). Unity doesn't like that, so I deleted the duplicate in the webp package. Build was successful but if there are issues down the line at least we'll know why.