WIP: web asset cache
All files are now able to be downloaded and cached locally. "Connection closed unexpectedly" errors still potentially happen, but by spacing the creation of web requests out and limiting the number of active requests the issue doesn't seem to happen anymore (but downloading takes slightly longer as a result). File I/O is still synchronous so the client freezes while caching and loading, but it works. Added an animated "activity icon" to the test scene to see client responsiveness.