rewrote yarnspinner commands, updated demo conversation
Made the exposed yarn commands more consistent with the Yarn Spinner documentation and built in commands. Also rewrote the background related commands so the background object is no longer needed for the command. Updated the demo conversation to use the new commands, and tested some of the features Yarn Spinner comes packaged with.