fixed automatic dialogue start in cutscene
If the Yarn Spinner dialogue system is set to start automatically, there's a good chance that the other gameobjects in the scene haven't been spawned and/or initialized yet. That's what was happening before, so any commands at the start of a yarn script would usually fail to execute because the gameobjects they were manipulating didn't exist yet.

This change makes sure the dialogue system doesn't start until all objects have been initialized, so hopefully no errors occur and writers can use the start of a Yarn script to do necessary setup of the cutscene before the first line of dialogue appears.