renamed 'conversation' classes and references to 'cutscene' #16
Everything in the cutscene scene was named according to conversations and Yarn Spinner demos. This was a confusing reference to the scene's initial purpose as a simple demo; now that it's a full fledged cutscene I renamed everything to reflect that, so relevant code will be easier to find and read.