github workflows build only on dispatch
When I committed the last pull request I noticed that like 7 builds were being requested at the same time. Not only is this entirely unnecessary for checking if the branch is okay, it's a waste of time and limited bandwidth. The ubuntu runner variants of the full build and cutscene only build should produce working clients, and take a fraction of the time to complete. Those are the two I'm assigning to pull requests going forward.