rough gm screen skeleton
Added a minimal gm screen with a list of commands and as much information as I can provide about their usage. I'm planning on replacing this with a nicer version where the command information is displayed in a separate window to the right, but this will have to do for the moment.